uvm_sequencer_base: refactor print() function
Clang reported the hidden overloaded virtual function `sc_object::print`
in the class `uvm_sequencer_base`.  The `print` function is marked as
implementation-defined and private in this class.

To avoid the conflict with the SystemC function, this patch removes
the `print` function from `uvm_sequencer_base` and adds a `using`
declaration to the `uvm_component` base class instead, where the
ambiguity between SystemC and UVM is introduced first.